/seminar 리팩토링 및 서버 연결
