+++ apis/notice.ts
@@ -10,7 +10,7 @@ import { deleteRequest, getRequest, patchRequest, postRequest } from '.';
 const noticePath = '/notice';
 
 export const getNoticePosts = (params: PostSearchQueryParams) =>
-  getRequest(noticePath, params, { cache: 'no-store' }) as Promise<GETNoticePostsResponse>;
+  getRequest(noticePath, params, {cache: 'no-store'}) as Promise<GETNoticePostsResponse>;
 
 export const getNoticePostDetail = (id: number, params: PostSearchQueryParams) =>
   getRequest(`${noticePath}/${id}`, params) as Promise<NoticePostResponse>;
